chore: refactoring version page
