+++ src/components/Engines/Engines.tsx
@@ -5,7 +5,8 @@ import Grid from '@material-ui/core/Grid';
 import List from '@material-ui/core/List';
 import ListItemText from '@material-ui/core/ListItemText';
 
-import { DetailContextConsumer, VersionPageConsumerProps } from '../../pages/version/Version';
+import { VersionPageConsumerProps } from '../../pages/version/types';
+import { DetailContextConsumer } from '../../pages/version/context';
 import { Heading, EngineListItem } from './styles';
 // @ts-ignore
 import node from './img/node.png';
