+++ src/components/UpLinks/UpLinks.tsx
@@ -2,7 +2,7 @@ import React, { ReactElement } from 'react';
 import List from '@material-ui/core/List';
 import ListItem from '@material-ui/core/ListItem';
 
-import { DetailContextConsumer } from '../../pages/version/Version';
+import { DetailContextConsumer } from '../../pages/version/context';
 import NoItems from '../NoItems';
 import { formatDateDistance } from '../../utils/package';
